chore: DreamsNetworkInteracting to separate file
